requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
appdirs==1.4.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chumpy==0.70
click==8.1.8
cmake==3.31.6
colorama==0.4.6
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.1
filelock==3.14.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.1
json-tricks==3.17.3
kiwisolver==1.4.8
lit==18.1.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mat4py==0.6.0
matplotlib==3.10.1
mdurl==0.1.2
mmcv==2.1.0
mmdet==3.2.0
mmengine==0.10.7
mmpose==1.3.1
mmpretrain==1.2.0
model-index==0.1.11
modelindex==0.0.2
motmetrics==1.4.0
mpmath==1.3.0
munkres==1.1.4
networkx==3.4.2
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
protobuf==4.25.6
psutil==7.0.0
py-cpuinfo==9.0.0
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyTurboJPEG==1.8.2
pytz==2023.4
PyYAML==6.0.2
requests==2.28.2
rich==13.4.2
scikit-learn==1.7.0
scipy==1.15.2
sentry-sdk==2.23.1
setproctitle==1.3.5
shapely==2.0.7
six==1.17.0
smmap==5.0.2
sympy==1.13.3
tabulate==0.9.0
termcolor==2.5.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.2
triton==2.0.0
ttach==0.0.3
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.146
ultralytics-thop==2.0.14
urllib3==1.26.20
wandb==0.16.5
xmltodict==0.14.2
xtcocotools==1.14.3
yacs==0.1.8
yapf==0.43.0
zipp==3.21.0
